🚚 Move CEL rendering to engine/render/cel_render.cpp
