+++ Source/doom.cpp
@@ -7,6 +7,7 @@
 
 #include "control.h"
 #include "engine.h"
+#include "engine/render/cel_render.hpp"
 #include "utils/stdcompat/optional.hpp"
 
 namespace devilution {
