+++ Source/error.cpp
@@ -6,6 +6,7 @@
 #include "error.h"
 
 #include "control.h"
+#include "engine/render/cel_render.hpp"
 #include "stores.h"
 #include "utils/language.h"
